User documentation: including images is working again.
It was necessary to explicitly specify the "basedir" FOP Ant task parameter to include images again. Actually this should not be needed, as the FOP processor actually should lookup images relative to the FOP file – at least according to the FOP documentation as well as according to former experience with an explicit for.jar for FOP 2.1.

Now the FOP JAR is provided as a Maven dependency, but the version (2.1) is the same, so the behavior should actually be the same, too. Strange!