Change init and behavior of "Show Time Events" action
The "Show Time Events"-button in the simulation view was acting the exact
opposite of what a user would expect - the Tooltip says "Show Time Events",
but an activated button would make the time events disappear and vice versa.
The core reason for this is that the changed option is "HideTimeEvents".
If that is set to True, the time events are hidden, and the button is activated.
Now, the button is set to the opposite of the option so that it better fits the
given tooltip. A more lazy approach would have been to change the tooltip to
"Hide Time Events".
Furthermore, an action was put into place so that the default value of the
option is now "False" so that Time Events are displayed by default, making
the usage easier for beginners. Tiny buttons like these are easily overlooked.