Hotfix for validator malfunction
On a Linux machine running OpenJDK, Yakindu SCT validated History-states as
having a bad value in 'kind' (that is ENTRY, HISTORY or DEEP_HISTORY).
This could not be reproduced on a Windows machine.
For some reason, the validator runs into another function on Linux
that always returns false.
This was fixed by simply returning true instead.