Fix parameter compatibility validation
I found out that the TypeInferrer would only throw a warning when the user
tried to assign something like an integer to a parameter of type
ComplexType<T>, saying that T could not be inferred. This is of cause not the
real problem, as giving an integer to a ComplexType Parameter should not be
possible at all.
Rearranging some checks and exceptions fixed the problem.