#1323 : introduce Fast, Normal & Expensive SCT markers
was found during testing of #1305, where markers from normal and
expensive checks get deleted when live validation occurs.

this was because the information of Xtext check type got lost when
markers got created(SCTMarkerTypeProvider), but the validation store
relies on the Xtext check type to ensure all errors are consistently
shown.

see also #1305