Enforce minimum height of documentation property section (#1335)
This change sets the minimum height of the documentation text field in
the properties to be 3 times the height of a single line. In practice at
least on my windows machine still only 2 lines (and some empty pixels)
are visible at any time.

If this is not done, the text control resizes up to no height at all, if
the remaining controls (like list of transitions in the state) gets
long.