Fixes #1800. For subdiagrams the definition section is only visible i… (#1818)
* Fixes #1800. For subdiagrams the definition section is only visible if the pinning feature is activated.
Fixes #1799. The definition section now really remembers the last state of expansion (also for subdiagrams).
Fixes #1798. Replaced the SWT button control with SWT label to eliminate size issues of the switching control in the statechart definition section on different platforms.

For the pinnedn definition section the name at the top is now always the name of the root diagram.

* Simplified getter for statechart name for definition section