Remove IdentifierSettings "maxLength" and "statechartPrefix"
Everything related to name shortening will be moved to the Professional Edition
as this is a feature mainly needed for professional users that want to comply
with certain coding rules. Instead of shortening the names magically, the
statechart will issue warnings for names that are too long.
Therefore, the statemachine prefix has to be removed, because the statechart
has no knowledge of the sgen or options used therein.
The setting maxIdentifierLength is removed. Instead an annotation will be
introduced, only available in C Domain Statecharts, which allows the user
to specify he doesn't want any names longer than 31 chars.