Remove ISCTExtension (#2110)
Removed the Interface ISCTExtension, because its obsolete and the class
STextExtensions alone is enough. Therefore replaced the injections of
the interface with the injections of the classes.
Despite this, the method "getScopeElement" was extracted from the
STextExtensions to the STextScopeProvider because this was the only
place where the methods was used. Also a getStatechart wrapper method
was recreated in the STextScopeProvider, so that this method can be
overridden in the SCTUnitScopeProvider.

This fixes #2110