Initial SText quickfix implementation for unknown events (#2738)
Also, a serializer bug has been fixed, which only occoured while
testing the SText Quickfixes. Given a Definition-Section containing only
the default annotations, the quickfix for creating both the event and
its corresponding unnamned interface would eat up all spaces between the
Annotations and the newly created interface.

That caused syntactic failures, as the Annotation and the interface
could not be correctly parsed anymore - no whitespace was between the
keywords.