Execute timer task after re-scheduling it to ensure periodic ta… (#3024)
In event-driven mode, the timer task callback might unschedule the timer task itself, e.g. on a state change, but with the old code, it will be re-scheduled again. This leads to strange behavior for "every" time triggers (see issue #3023). Therefore, we need to run the task after it was re-scheduled, so it can be properly unscheduled by its callback.